@@ -85,7 +85,7 @@ class FancyCard extends StatelessWidget {
             ),
             Text(
               title,
-              style: Theme.of(context).textTheme.headline5,
+              style: Theme.of(context).textTheme.headlineSmall,
             ),
             OutlinedButton(
               child: const Text("Learn more"),
